Refactored Code
Updated controller to use more logical endpoint structure and removed logic from the controller.
Update dataservice class to purely retrieve and modify data in a collection and added 2 extensions to this class that can be used for local development and production environment.
Added client for CurrencyLayer api and moved Apikey to appsettings.
Added handlers for profiles and currency rates.
Added tests for the 3 core functions of the api, retrieving a portfolio, retrieveng the value of a portfolio and deleteing a portfolio.